requirements.txt
langchain
langgraph
langchain-google-genai
google-generativeai
faiss-cpu
sentence-transformers
pandas
python-dotenv
langchain-huggingface
